+++ src/test/java/woowacourse/auth/acceptance/AuthAcceptanceTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.http.HttpStatus;
-import woowacourse.auth.dto.TokenResponse;
+import woowacourse.auth.application.dto.TokenResponse;
 import woowacourse.shoppingcart.acceptance.AcceptanceTest;
 import woowacourse.shoppingcart.acceptance.ResponseCreator;
