+++ src/test/java/woowacourse/shoppingcart/acceptance/ResponseCreator.java
@@ -4,7 +4,7 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.springframework.http.MediaType;
-import woowacourse.auth.dto.TokenResponse;
+import woowacourse.auth.application.dto.TokenResponse;
 import woowacourse.auth.ui.dto.TokenRequest;
 import woowacourse.shoppingcart.ui.dto.CustomerChangePasswordRequest;
 import woowacourse.shoppingcart.ui.dto.CustomerSignUpRequest;
